+++ packages/erd-editor/src/engine/modules/table-column/hooks.ts
@@ -91,12 +91,30 @@ const validationForeignKeyHook: CO = function* (channel, state) {
     const relationships = query(collections)
       .collection('relationshipEntities')
       .selectByIds(doc.relationshipIds);
+    const tables = query(collections)
+      .collection('tableEntities')
+      .selectByIds(doc.tableIds);
+    const foreignKeyColumnIdsSet = new Set<string>();
 
     for (const { end } of relationships) {
       query(collections)
         .collection('tableColumnEntities')
         .updateMany(end.columnIds, column => {
           column.ui.keys = column.ui.keys | ColumnUIKey.foreignKey;
+          foreignKeyColumnIdsSet.add(column.id);
+        });
+    }
+
+    for (const table of tables) {
+      query(collections)
+        .collection('tableColumnEntities')
+        .updateMany(table.columnIds, column => {
+          if (
+            bHas(column.ui.keys, ColumnUIKey.foreignKey) &&
+            !foreignKeyColumnIdsSet.has(column.id)
+          ) {
+            column.ui.keys = column.ui.keys & ~ColumnUIKey.foreignKey;
+          }
         });
     }
   });