+++ packages/erd-editor-vscode/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 1.0.14
+
+- Validation foreignKey
+
 ## 1.0.11
 
-- legacy back porting
+- Legacy back porting
 
 ## 1.0.9